refactor(infrastructure): split GeologyLayer into focused services
Extract GeologyLayer monolith (432 lines) into 5 specialized services
following the same orchestrator pattern as TopographyLayer, HydrologyLayer,
VegetationLayer, and StructuresLayer.

New services in packages/infrastructure/src/map/services/layers/geology/:
- FormationSelectionService: selectFormations() + getFormationsForBiome()
  + createSeededRandom() — biome-to-formation mapping and seeded selection
- BedrockPatternService: generatePattern() — noise-based 2D formation grid
  with bedding-orientation structural patterns
- WeatheringService: applyWeathering() — intensity-driven feature selection
  (major/intermediate/depression features per rock type)
- SoilCalculationService: calculateDepths() — feature-modified soil depth
  calculation with noise variation
- GeologyTileGenerationService: createTiles() + findTransitionZones()
  — final tile assembly and formation boundary detection

GeologyLayer becomes a lightweight orchestrator (~100 lines):
- Injects 5 services via class-based @Inject() tokens (no string tokens)
- @Optional() ILogger for structured logging
- Width/height passed as parameters to each service (no instance state)

Register 5 new providers in infrastructure.module.ts under
'Geology internal services' section.
